rag_research/requirements.txt
fastapi
uvicorn
streamlit
pymupdf
openai
requests
python-multipart
python-dotenv
numpy
